Add README files on repo and directory

+++ 0x00-python_variable_annotations/0-add.py
@@ -1,6 +1,7 @@
-#!/usr/bin/python3
+#!/usr/bin/env python3
 """Module defines type-annotated `add` function"""
 
-def add(a: int, b: float) -> float:
+
+def add(a: float, b: float) -> float:
     """Return sum of two floats"""
     return a + b
